requirements.txt
pandas
fastapi
uvicorn
python-multipart
pytest
pytest-asyncio
pytest-cov
httpx
psycopg2-binary
sqlalchemy